pr 요청 드립니다.

+++ src/main/java/com/app/project/domain/PostDTO.java
@@ -13,7 +13,7 @@ public class PostDTO extends Period {
     private Long id;
     private String postTitle;
     private String postContent;
-    private Integer viewCount;
+    private Long viewCount;
     private Long memberId;
     private String memberEmail;
     private String memberName;
